src/main/java/com/example/geosuggest/suggestion/SuggestionDTO.java
package com.example.geosuggest.suggestion;

import java.io.Serializable;

public class SuggestionDTO implements Serializable {
    private String name;
    private double latitude;
    private double longitude;
    private double score;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public double getLatitude() { return latitude; }

    public void setLatitude(double latitude) {
        this.latitude = latitude;
    }

    public double getLongitude() {
        return longitude;
    }

    public void setLongitude(double longitude) {
        this.longitude = longitude;
    }

    public double getScore() {
        return score;
    }

    public void setScore(double score) {
        this.score = score;
    }
}
